requirements.txt
# Core Framework
streamlit>=1.28.0
streamlit-aggrid>=0.3.4

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# Document Processing
python-docx>=0.8.11

# Git Operations
GitPython>=3.1.30

# HTTP Requests
requests>=2.31.0

# File Operations
pathlib>=1.0.1

# Logging
loguru>=0.7.0

# Progress Bars
tqdm>=4.65.0

# Visualization
plotly>=5.15.0

# Functional Testing
playwright>=1.40.0
selenium>=4.15.0
beautifulsoup4>=4.12.0

# Future AI Integration
# semantic-kernel>=0.5.0  # Uncomment when implementing Phase 2
